backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14
django-cors-headers>=4.0
python-dotenv>=1.0

# Database
psycopg2-binary>=2.9

# Production server
gunicorn>=21.0
whitenoise>=6.6

# HTTP requests (for Telegram, TronGrid)
requests>=2.31
httpx>=0.25

# Crypto
cryptography>=41.0

# Async support
aiohttp>=3.9
